refactor: 코드 폴더 구조 리팩토링(로직 별로)
